refactor: update loaders for integer positions and removal of robot radius
- load_robot_states: cast position x/y to int on load (handles legacy float JSON)
- load_robots: speed must be a positive int; remove radius field and validation

+++ scenario_loaders/load_robot_states.py
@@ -67,7 +67,7 @@ def load_robot_states(raw: list[dict[str, Any]]) -> list[RobotState]:
 
         robot_state = RobotState(
             robot_id=RobotId(robot_id),
-            position=Position(float(x), float(y)),
+            position=Position(int(x), int(y)),
             battery_level=float(battery_level),
         )
         robot_states.append(robot_state)
